Official SDL Binding
The SDL binding still has some limitation in the events implmenent but mouse/keyboard events has been added as those are the most important events right now.

All other features are the same across the SDL binding as with the SFML binding.

The engine will run and render in the same way using both of them.